requirements.txt
click==8.0.3
# dataclasses==0.8
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==2.11.3
MarkupSafe==2.0.1
pyodbc==4.0.32
python-dotenv==0.19.2
typing_extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
